+++ Scripts/Save/GameSave.cs
@@ -4,7 +4,7 @@ public class GameSave {
     public string Name = "Company Inc";
     public string Ceo = "Gill Bates";
     public string[] Industries = {"making money"};
-    public int Money = 50000;
+    public long Money = 50000;
     public int Year = 1;
     public int Month = 1;
     public int Week = 1;
